changed a small line in testcontroller
created a BaseEntity with all the createdOn and UpdatedOn attributes to be inherited in other models

Removed createdby and updatedBy etc in all the models and refrenced to BaseEntity.

Added binding tables: UserGroupMembers and UserGroupResources

Cleaned up DBContext and added an override on the savechanges() to set updatedOn and CreatedOn if there are changes to database content.

Added a method in SeedData that is supposed to create data in table. (not finished)